Type Racket solution to multikata square list (#5)
Show off the ability of Racket to be multiple languages at the same time by
typing only the part of the solution module that holds the `comp` function.
The tests are still untyped Racket code. There is a new test-case that checks
that now only lists of integer numbers can cross the typed/untyped boundary
without causing a contract exception.

The internal definition of `sqr` is now in a `define` form instead of a lambda
receiving a name within a `let`. The Racket compiler did not behave well with
just a type annotation on top of `comp`, but probably the unchanged code would
have worked with some more type hints within its body.